Move copy command (#1677)
* Move copy command

* Remove unused using directives
